backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-django>=4.11.1",
    "ruff>=0.14.6",
]
